+++ omisego-core/src/main/java/co/omisego/omisego/model/TransactionRequest.kt
@@ -1,4 +1,4 @@
-package co.omisego.omisego.model.transaction.request
+package co.omisego.omisego.model
 
 /*
  * OmiseGO
@@ -9,11 +9,7 @@ package co.omisego.omisego.model.transaction.request
 
 import android.os.Parcelable
 import co.omisego.omisego.constant.enums.OMGEnum
-import co.omisego.omisego.model.Account
-import co.omisego.omisego.model.Token
-import co.omisego.omisego.model.User
 import co.omisego.omisego.model.socket.SocketTopic
-import co.omisego.omisego.model.transaction.consumption.TransactionConsumptionParams
 import co.omisego.omisego.operation.Listenable
 import kotlinx.android.parcel.Parcelize
 import kotlinx.android.parcel.RawValue
@@ -102,35 +98,3 @@ data class TransactionRequest(
     val metadata: @RawValue Map<String, Any>,
     val encryptedMetadata: @RawValue Map<String, Any>
 ) : Parcelable, Listenable
-
-/**
- * An extension function that converts the [TransactionRequest] to the [TransactionConsumptionParams] easily
- *
- * @param amount The amount of token to transfer (down to subunit to unit)
- * @param address The address to use for the consumption
- * In the case of a type "send", this will be the token that the consumer will receive
- * In the case of a type "receive" this will be the token that the consumer will send
- * @param idempotencyToken The idempotency token to use for the consumption
- * @param correlationId An id that can uniquely identify a transaction. Typically an order id from a provider.
- * @param metadata Additional metadata for the consumption
- * @param encryptedMetadata Additional encrypted metadata for the consumption
- *
- * @return The [TransactionConsumptionParams] used for consume the a transaction request
- */
-fun TransactionRequest.toTransactionConsumptionParams(
-    amount: BigDecimal? = null,
-    address: String? = null,
-    idempotencyToken: String = "${this.id}-${System.nanoTime()}",
-    correlationId: String? = null,
-    metadata: Map<String, Any> = mapOf(),
-    encryptedMetadata: Map<String, Any> = mapOf()
-): TransactionConsumptionParams =
-    TransactionConsumptionParams.create(
-        this,
-        amount,
-        address,
-        idempotencyToken,
-        correlationId,
-        metadata,
-        encryptedMetadata
-    )