+++ omisego-core/src/test/kotlin/co/omisego/omisego/operation/ListenableTest.kt
@@ -10,8 +10,8 @@ package co.omisego.omisego.operation
 import co.omisego.omisego.model.APIError
 import co.omisego.omisego.model.User
 import co.omisego.omisego.model.socket.SocketTopic
-import co.omisego.omisego.model.transaction.consumption.TransactionConsumption
-import co.omisego.omisego.model.transaction.request.TransactionRequest
+import co.omisego.omisego.model.TransactionConsumption
+import co.omisego.omisego.model.TransactionRequest
 import co.omisego.omisego.websocket.SocketClientContract
 import co.omisego.omisego.websocket.event.TransactionConsumptionRequestEvent
 import co.omisego.omisego.websocket.listener.DelegateSocketCustomEventListener