+++ omisego-core/src/test/kotlin/co/omisego/omisego/websocket/strategy/FilterStrategyTest.kt
@@ -9,7 +9,7 @@ package co.omisego.omisego.websocket.strategy
 
 import co.omisego.omisego.model.socket.SocketReceive
 import co.omisego.omisego.model.socket.SocketTopic
-import co.omisego.omisego.model.transaction.consumption.TransactionConsumption
+import co.omisego.omisego.model.TransactionConsumption
 import co.omisego.omisego.utils.Either
 import co.omisego.omisego.websocket.enum.SocketCustomEvent
 import co.omisego.omisego.websocket.event.SocketEvent